Revert any changes having side effects on code generation
