+++ Core/GDCore/IDE/Events/ExpressionVariableOwnerFinder.h
@@ -3,11 +3,11 @@
  * Copyright 2008-present Florian Rival (Florian.Rival@gmail.com). All rights
  * reserved. This project is released under the MIT License.
  */
-#pragma once
+#ifndef GDCORE_EXPRESSIONVARIABLEOWNERFINDER_H
+#define GDCORE_EXPRESSIONVARIABLEOWNERFINDER_H
 
 #include <memory>
 #include <vector>
-
 #include "GDCore/Events/Parsers/ExpressionParser2Node.h"
 #include "GDCore/Events/Parsers/ExpressionParser2NodeWorker.h"
 #include "GDCore/Extensions/Metadata/ExpressionMetadata.h"
@@ -15,8 +15,7 @@
 #include "GDCore/Extensions/Metadata/ObjectMetadata.h"
 #include "GDCore/Extensions/Metadata/ParameterMetadata.h"
 #include "GDCore/Project/Layout.h"  // For GetTypeOfObject and GetTypeOfBehavior
-#include "GDCore/Project/ProjectScopedContainers.h"
-#include "GDCore/Project/Variable.h"
+#include "GDCore/Project/ObjectsContainersList.h"
 #include "GDCore/Tools/Localization.h"
 
 namespace gd {
@@ -29,65 +28,53 @@ class ExpressionMetadata;
 
 namespace gd {
 
-struct VariableOrVariablesContainer {
-  const gd::VariablesContainer* variablesContainer;
-  const gd::Variable* variable;
-};
-
 /**
  * \brief Find the object name that should be used as a context of the
- * expression or sub-expression that a given node represents, or find
- * the last parent of a node representing a variable.
+ * expression or sub-expression that a given node represents.
  *
- * Object name can be needed because of the legacy convention where a "objectvar"
- * parameter represents a variable of the object represented by the previous
- * "object" parameter.
+ * This is needed because of the legacy convention where a "objectvar"
+ * parameter represents a variable of the object represented by the previous "object"
+ * parameter.
  *
  * \see gd::ExpressionParser2
  */
-class GD_CORE_API ExpressionVariableOwnerFinder
-    : public ExpressionParser2NodeWorker {
+class GD_CORE_API ExpressionVariableOwnerFinder : public ExpressionParser2NodeWorker {
  public:
+
   /**
    * \brief Helper function to find the object name that should be used as a
    * context of the expression or sub-expression that a given node represents.
    */
-  static const gd::String GetObjectName(
-      const gd::Platform& platform,
-      const gd::ProjectScopedContainers& projectScopedContainers,
-      const gd::String& rootObjectName,
-      gd::ExpressionNode& node) {
+  static const gd::String GetObjectName(const gd::Platform &platform,
+                      const gd::ObjectsContainersList &objectsContainersList,
+                      const gd::String& rootObjectName,
+                      gd::ExpressionNode& node) {
     gd::ExpressionVariableOwnerFinder typeFinder(
-        platform, projectScopedContainers, rootObjectName);
+        platform, objectsContainersList, rootObjectName);
     node.Visit(typeFinder);
-    return typeFinder.objectName;
-  }
-
-  static VariableOrVariablesContainer GetLastParentOfNode(
-      const gd::Platform& platform,
-      const gd::ProjectScopedContainers& projectScopedContainers,
-      const gd::String& rootObjectName,
-      gd::ExpressionNode& node) {
-    gd::ExpressionVariableOwnerFinder typeFinder(
-        platform, projectScopedContainers, rootObjectName);
-    node.Visit(typeFinder);
-    return typeFinder.lastParentOfNode;
+    return typeFinder.GetObjectName();
   }
 
   virtual ~ExpressionVariableOwnerFinder(){};
 
  protected:
-  ExpressionVariableOwnerFinder(
-      const gd::Platform& platform_,
-      const gd::ProjectScopedContainers& projectScopedContainers_,
-      const gd::String& rootObjectName_)
+  ExpressionVariableOwnerFinder(const gd::Platform &platform_,
+                       const gd::ObjectsContainersList &objectsContainersList_,
+                       const gd::String& rootObjectName_)
       : platform(platform_),
-        projectScopedContainers(projectScopedContainers_),
-        objectName(),
+        objectsContainersList(objectsContainersList_),
         rootObjectName(rootObjectName_),
-        variableNode(nullptr),
-        thisIsALegacyPrescopedVariable(false),
-        legacyPrescopedVariablesContainer(nullptr){};
+        objectName(""),
+        variableNode(nullptr) {};
+
+  /**
+   * \brief Get all the errors
+   *
+   * No errors means that the expression is valid.
+   */
+  const gd::String &GetObjectName() {
+    return objectName;
+  };
 
   void OnVisitSubExpressionNode(SubExpressionNode& node) override {}
   void OnVisitOperatorNode(OperatorNode& node) override {}
@@ -99,151 +86,32 @@ class GD_CORE_API ExpressionVariableOwnerFinder
       // This is not possible
       return;
     }
-    variableNode = &node;
-
-    // Check if the parent is a function call, in which we might be dealing
-    // with a legacy pre-scoped variable parameter:
-    if (node.parent) node.parent->Visit(*this);
-    else {
+    if (node.parent == nullptr) {
       objectName = rootObjectName;
+      return;
     }
-
-    if (thisIsALegacyPrescopedVariable) {
-      // The node represents a variable name, and the variables container
-      // containing it was identified in the FunctionCallNode.
-      childVariableNames.insert(childVariableNames.begin(), node.name);
-      if (legacyPrescopedVariablesContainer)
-        lastParentOfNode = WalkUntilLastParent(
-            *legacyPrescopedVariablesContainer, childVariableNames);
-    } else {
-      // Otherwise, the identifier is to be interpreted as usual:
-      // it can be an object (on which a variable is accessed),
-      // or a variable.
-      projectScopedContainers.MatchIdentifierWithName<void>(
-          node.name,
-          [&]() {
-            // This is an object.
-            objectName = node.name;
-
-            const auto* variablesContainer =
-                projectScopedContainers.GetObjectsContainersList()
-                    .GetObjectOrGroupVariablesContainer(objectName);
-            if (variablesContainer)
-              lastParentOfNode =
-                  WalkUntilLastParent(*variablesContainer, childVariableNames);
-          },
-          [&]() {
-            // This is a variable.
-            if (projectScopedContainers.GetVariablesContainersList().Has(
-                    node.name)) {
-              lastParentOfNode = WalkUntilLastParent(
-                  projectScopedContainers.GetVariablesContainersList().Get(
-                      node.name),
-                  childVariableNames);
-            }
-          },
-          [&]() {
-            // Ignore properties here.
-            // There is no support for "children" of properties.
-          },
-          [&]() {
-            // Ignore parameters here.
-            // There is no support for "children" of parameters.
-          },
-          [&]() {
-            // Ignore unrecognised identifiers here.
-          });
-    }
+    variableNode = &node;
+    node.parent->Visit(*this);
   }
-  void OnVisitVariableAccessorNode(VariableAccessorNode& node) override {
-    childVariableNames.insert(childVariableNames.begin(), node.name);
-    if (node.parent) node.parent->Visit(*this);
-  }
+  void OnVisitVariableAccessorNode(VariableAccessorNode& node) override {}
   void OnVisitIdentifierNode(IdentifierNode& node) override {
     if (variableNode != nullptr) {
       // This is not possible
       return;
     }
-    // This node is not necessarily a variable node.
-    // It will be checked when visiting the FunctionCallNode, just after.
-    variableNode = &node;
-
-    // Check if the parent is a function call, in which we might be dealing
-    // with a legacy pre-scoped variable parameter:
-    if (node.parent) node.parent->Visit(*this);
-    else {
+    if (node.parent == nullptr) {
       objectName = rootObjectName;
+      return;
     }
-
-    if (thisIsALegacyPrescopedVariable) {
-      // The identifier represents a variable name, and the variables container
-      // containing it was identified in the FunctionCallNode.
-      if (!node.childIdentifierName.empty())
-        childVariableNames.insert(childVariableNames.begin(),
-                                  node.childIdentifierName);
-      childVariableNames.insert(childVariableNames.begin(),
-                                node.identifierName);
-
-      if (legacyPrescopedVariablesContainer)
-        lastParentOfNode = WalkUntilLastParent(
-            *legacyPrescopedVariablesContainer, childVariableNames);
-
-    } else {
-      // Otherwise, the identifier is to be interpreted as usual:
-      // it can be an object (on which a variable is accessed),
-      // or a variable.
-      projectScopedContainers.MatchIdentifierWithName<void>(
-          node.identifierName,
-          [&]() {
-            // This is an object.
-            objectName = node.identifierName;
-            if (!node.childIdentifierName.empty())
-              childVariableNames.insert(childVariableNames.begin(),
-                                        node.childIdentifierName);
-
-            const auto* variablesContainer =
-                projectScopedContainers.GetObjectsContainersList()
-                    .GetObjectOrGroupVariablesContainer(objectName);
-            if (variablesContainer)
-              lastParentOfNode =
-                  WalkUntilLastParent(*variablesContainer, childVariableNames);
-          },
-          [&]() {
-            // This is a variable.
-            if (!node.childIdentifierName.empty())
-              childVariableNames.insert(childVariableNames.begin(),
-                                        node.childIdentifierName);
-
-            if (projectScopedContainers.GetVariablesContainersList().Has(
-                    node.identifierName)) {
-              lastParentOfNode = WalkUntilLastParent(
-                  projectScopedContainers.GetVariablesContainersList().Get(
-                      node.identifierName),
-                  childVariableNames);
-            }
-          },
-          [&]() {
-            // Ignore properties here.
-            // There is no support for "children" of properties.
-          },
-          [&]() {
-            // Ignore parameters here.
-            // There is no support for "children" of properties.
-          },
-          [&]() {
-            // Ignore unrecognised identifiers here.
-          });
-    }
+    // This node is not necessarily a variable node.
+    // It will be checked when visiting the FunctionCallNode.
+    variableNode = &node;
+    node.parent->Visit(*this);
   }
   void OnVisitEmptyNode(EmptyNode& node) override {}
   void OnVisitObjectFunctionNameNode(ObjectFunctionNameNode& node) override {}
   void OnVisitVariableBracketAccessorNode(
-      VariableBracketAccessorNode& node) override {
-    // Add a child with an empty name, which will be interpreted as
-    // "take the first child/item of the structure/array".
-    childVariableNames.insert(childVariableNames.begin(), "");
-    if (node.parent) node.parent->Visit(*this);
-  }
+      VariableBracketAccessorNode& node) override {}
   void OnVisitFunctionCallNode(FunctionCallNode& functionCall) override {
     if (variableNode == nullptr) {
       return;
@@ -258,130 +126,46 @@ class GD_CORE_API ExpressionVariableOwnerFinder
     if (parameterIndex < 0) {
       return;
     }
-
-    const auto& objectsContainersList =
-        projectScopedContainers.GetObjectsContainersList();
-
     const gd::ParameterMetadata* parameterMetadata =
         MetadataProvider::GetFunctionCallParameterMetadata(
-            platform, objectsContainersList, functionCall, parameterIndex);
-    if (parameterMetadata == nullptr) return;  // Unexpected
+            platform,
+            objectsContainersList,
+            functionCall,
+            parameterIndex);
+    if (parameterMetadata == nullptr
+     || parameterMetadata->GetType() != "objectvar") {
+      return;
+    }
 
-    // Support for legacy pre-scoped variables:
-    if (parameterMetadata->GetValueTypeMetadata().IsLegacyPreScopedVariable()) {
-      if (parameterMetadata->GetType() == "objectvar") {
-        // Legacy convention where a "objectvar"
-        // parameter represents a variable of the object represented by the
-        // previous "object" parameter. The object on which the function is
-        // called is returned if no previous parameters are objects.
-        objectName = functionCall.objectName;
-        for (int previousIndex = parameterIndex - 1; previousIndex >= 0;
-             previousIndex--) {
-          const gd::ParameterMetadata* previousParameterMetadata =
-              MetadataProvider::GetFunctionCallParameterMetadata(
-                  platform, objectsContainersList, functionCall, previousIndex);
-          if (previousParameterMetadata != nullptr &&
-              gd::ParameterMetadata::IsObject(
-                  previousParameterMetadata->GetType())) {
-            auto previousParameterNode =
-                functionCall.parameters[previousIndex].get();
-            IdentifierNode* objectNode =
-                dynamic_cast<IdentifierNode*>(previousParameterNode);
-            objectName = objectNode->identifierName;
-            break;
-          }
-        }
-
-        legacyPrescopedVariablesContainer =
-            projectScopedContainers.GetObjectsContainersList()
-                .GetObjectOrGroupVariablesContainer(objectName);
-        thisIsALegacyPrescopedVariable = true;
-      } else if (parameterMetadata->GetType() == "scenevar") {
-        legacyPrescopedVariablesContainer =
-            projectScopedContainers.GetVariablesContainersList()
-                .GetBottomMostVariablesContainer();
-        thisIsALegacyPrescopedVariable = true;
-      } else if (parameterMetadata->GetType() == "globalvar") {
-        legacyPrescopedVariablesContainer =
-            projectScopedContainers.GetVariablesContainersList()
-                .GetTopMostVariablesContainer();
-        thisIsALegacyPrescopedVariable = true;
+    // The object on which the function is called is returned if no previous
+    // parameters are objects.
+    objectName = functionCall.objectName;
+    for (int previousIndex = parameterIndex - 1; previousIndex >= 0; previousIndex--) {
+      const gd::ParameterMetadata* previousParameterMetadata =
+          MetadataProvider::GetFunctionCallParameterMetadata(
+              platform,
+              objectsContainersList,
+              functionCall,
+              previousIndex);
+      if (previousParameterMetadata != nullptr
+      && gd::ParameterMetadata::IsObject(previousParameterMetadata->GetType())) {
+        auto previousParameterNode = functionCall.parameters[previousIndex].get();
+        IdentifierNode* objectNode = dynamic_cast<IdentifierNode*>(previousParameterNode);
+        objectName = objectNode->identifierName;
+        return;
       }
-    } else {
-      thisIsALegacyPrescopedVariable = false;
-      legacyPrescopedVariablesContainer = nullptr;
     }
   }
 
  private:
-  static VariableOrVariablesContainer WalkUntilLastParent(
-      const gd::Variable& variable,
-      const std::vector<gd::String>& childVariableNames,
-      size_t startIndex = 0) {
-    const gd::Variable* currentVariable = &variable;
-
-    // Walk until size - 1 as we want the last parent.
-    for (size_t index = startIndex; index + 1 < childVariableNames.size();
-         ++index) {
-      const gd::String& childName = childVariableNames[index];
-
-      if (childName.empty()) {
-        if (currentVariable->GetChildrenCount() == 0) {
-          // The array or structure is empty, we can't walk through it - there
-          // is no "parent".
-          return {};
-        }
-
-        if (currentVariable->GetType() == gd::Variable::Array) {
-          currentVariable = &currentVariable->GetAtIndex(0);
-        } else {
-          currentVariable =
-              currentVariable->GetAllChildren().begin()->second.get();
-        }
-      } else {
-        if (!currentVariable->HasChild(childName)) {
-          // Non existing child - there is no "parent".
-          return {};
-        }
-
-        currentVariable = &currentVariable->GetChild(childName);
-      }
-    }
-
-    // Return the last parent of the chain of variables (so not the last variable
-    // but the one before it).
-    return {.variable = currentVariable};
-  }
-
-  static VariableOrVariablesContainer WalkUntilLastParent(
-      const gd::VariablesContainer& variablesContainer,
-      const std::vector<gd::String>& childVariableNames) {
-    if (childVariableNames.empty())
-      return {};  // There is no "parent" to the variables container itself.
-    if (childVariableNames.size() == 1)
-      return {// Only one child: the parent is the variables container itself.
-              .variablesContainer = &variablesContainer};
-
-    const gd::String& firstChildName = *childVariableNames.begin();
-    if (!variablesContainer.Has(firstChildName)) {
-      // The child does not exist - there is no "parent".
-      return {};
-    }
-
-    return WalkUntilLastParent(
-        variablesContainer.Get(firstChildName), childVariableNames, 1);
-  }
-
   gd::String objectName;
-  const gd::String& rootObjectName;
-  gd::ExpressionNode* variableNode;
-  std::vector<gd::String> childVariableNames;
-  bool thisIsALegacyPrescopedVariable;
-  const gd::VariablesContainer* legacyPrescopedVariablesContainer;
-  VariableOrVariablesContainer lastParentOfNode;
+  gd::ExpressionNode *variableNode;
 
-  const gd::Platform& platform;
-  const gd::ProjectScopedContainers& projectScopedContainers;
+  const gd::Platform &platform;
+  const gd::ObjectsContainersList &objectsContainersList;
+  const gd::String &rootObjectName;
 };
 
 }  // namespace gd
+
+#endif  // GDCORE_EXPRESSIONVARIABLEOWNERFINDER_H
